requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0